requirements.txt
# Web Framework & Server
flask==2.3.3
gunicorn==21.2.0
flask-cors==4.0.0
werkzeug==2.3.7

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0

# Machine Learning & Spam Detection
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
joblib==1.3.2
nltk==3.8.1

# Text Processing
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# OCR & Image Processing  
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
Pillow==10.1.0
easyocr==1.7.0

# Database
psycopg2-binary==2.9.9

# HTTP & Requests
requests==2.31.0
urllib3==2.1.0

# Email Processing
email-validator==2.1.0

# File Processing
python-multipart==0.0.6

# Utilities
tqdm==4.66.1
six==1.16.0
